Change format video step to merge audio and video
Previously I have tried combining the audio with the video at the very end, but this leads to the problem where the audio and video are out of sync. Instead now, we can combine the audio and video before the splitting to clip step occurs. Hopefully this will improve audio quality while ensuring that nothing gets out of sync.